re-factor: prepare class and pkg names for moving to server module
